Implemented some of the feedback
